Moving actor methods out of utils
