+++ module/helpers/utils.mjs
@@ -64,22 +64,6 @@ export async function createItemCopies(items, owner, type, parentItem, lastProce
 }
 
 /**
- * Handle looking up tokens associated with actor and changing size
- * @param {Actor} actor  The actor
- * @param {Number} width The actor's new width
- * @param {Number} height The actor's new width
- */
-export function resizeTokens(actor, width, height) {
-  const tokens = actor.getActiveTokens();
-  for (const token of tokens) {
-    token.document.update({
-      "height": height,
-      "width": width,
-    });
-  }
-}
-
-/**
  * Handle Shifting skills
  * @param {String} skill The skill shifting
  * @param {Number} shift The quantity of the shift
@@ -107,19 +91,6 @@ export function getShiftedSkill(skill, shift, actor) {
   }
 
   return [newShift, skillString];
-}
-
-/**
- * Displays an error message if the sheet is locked
- * @returns {boolean} True if the sheet is locked, and false otherwise
- */
-export function checkIsLocked(actor) {
-  if (actor.system.isLocked) {
-    ui.notifications.error(game.i18n.localize('E20.ActorLockError'));
-    return true;
-  }
-
-  return false;
 }
 
 /**
@@ -452,19 +423,4 @@ export async function setFocusValues(focus, actor, newLevel=null, previousLevel=
     // Level down
     await deleteAttachmentsForItem(focus, actor, previousLevel);
   }
-}
-
-/**
- * Prepare the number of actions available for the given actor
- * @param {Actor} actor The actor to get actions for
- * @return {Object} Action types mapped to an action count
- */
-export function getNumActions(actor) {
-  const speed = actor.system.essences.speed;
-
-  return {
-    free: Math.max(0, speed - 2),
-    movement: speed > 0 ? 1 : 0,
-    standard: speed > 1 ? 1 : 0,
-  };
 }
